+++ aerospike-core/src/policy/mod.rs
@@ -40,6 +40,8 @@ pub use self::admin_policy::AdminPolicy;
 pub use self::batch_policy::BatchPolicy;
 pub use self::client_policy::AuthMode;
 pub use self::client_policy::ClientPolicy;
+#[cfg(feature = "tls")]
+pub use self::client_policy::TlsPolicy;
 pub use self::client_policy::TEND_INTERVAL_MIN_MS;
 pub use self::commit_level::CommitLevel;
 pub use self::concurrency::Concurrency;